LPS-178068 Create WebClientConfiguration
